# make data model compatible with metricbeat