--- a/CHANGELOG.asciidoc
+++ b/CHANGELOG.asciidoc
@@ -20,6 +20,7 @@ https://github.com/elastic/apm-server/compare/6.3\...master[View commits]
 - Add /v1/metrics endpoint {pull}1000[1000].
 - Remove regexProperties validation rules {pull}1148[1148], {pull}1150[1150].
 - Add source_mapping.elasticsearch configuration option {pull}1114[1114].
+- Add /v1/metrics endpoint {pull}1000[1000] {pull}1121[1121].
 
 
 [[release-notes-6.3]]