--- a/model/metric/counter.go
+++ b/model/metric/counter.go
@@ -18,7 +18,6 @@
 package metric
 
 import (
-	"github.com/elastic/apm-server/utility"
 	"github.com/elastic/beats/libbeat/common"
 )
 
@@ -36,16 +35,7 @@ func (md *metricDecoder) decodeCounter(name string, raw map[string]interface{}) 
 	}
 }
 
-func (c *counter) mapstr() common.MapStr {
-	v := common.MapStr{
-		"type":  "counter",
-		"value": c.count,
-	}
-	utility.Add(v, "unit", c.unit)
-	return v
-}
-
 func (c *counter) transform(m common.MapStr) error {
-	m[c.name] = c.mapstr()
+	m.Put(c.name, c.count)
 	return nil
 }