--- a/model/metric/generated/schema/payload.go
+++ b/model/metric/generated/schema/payload.go
@@ -41,73 +41,17 @@ const PayloadSchema = `{
     "$id": "docs/spec/metrics/sample.json",
     "type": ["object", "null"],
     "description": "A single metric sample.",
-    "anyOf": [
-        {
-            "properties": {
-                "type": {
-                    "description": "Counters and gauges capture a single value at a point in time.  Counter are cumulative, strictly increasing or decreasing, and typically most useful with derivative aggregations.  Gauges increase and decrease over time.",
-                    "enum": ["counter", "gauge"]
-                },
-                "unit": {
-                    "type": ["string", "null"]
-                },
-                "value": {"type": "number"}
-            },
-            "required": ["type", "value"]
+    "properties": {
+        "type": {
+            "description": "Counters and gauges capture a single value at a point in time.  Counter are cumulative, strictly increasing or decreasing, and typically most useful with derivative aggregations.  Gauges increase and decrease over time.",
+            "enum": ["counter", "gauge"]
         },
-        {
-            "properties": {
-                "type": {
-                    "description": "Summary metrics capture client-side aggregations describing the distribution of a metric",
-                    "enum": ["summary"]
-                },
-                "unit": {
-                    "description": "The unit of measurement of this metric eg: bytes. Only informational at this time",
-                    "type": ["string", "null"]
-                },
-                "count": {
-                    "description": "The total count of all observations for this metric",
-                    "type": "number"
-                },
-                "sum": {
-                    "description": "The sum of all observations for this metric",
-                    "type": "number"
-                },
-                "stddev": {
-                    "description": "The standard deviation describing this metric",
-                    "type": ["number", "null"]
-                },
-                "min": {
-                    "description": "The minimum value observed for this metric",
-                    "type": ["number", "null"]
-                },
-                "max": {
-                    "description": "The maximum value observed for this metric",
-                    "type": ["number", "null"]
-                },
-                "quantiles": {
-                    "description": "A list of quantiles describing the metric",
-                    "type": ["array", "null"],
-                    "items": {
-                        "descrption": "A [quantile, value] tuple",
-                        "type": ["array", "null"],
-                        "items": [
-                            {
-                                "type": "number",
-                                "minimum": 0, "maximum": 1
-                            },
-                            {
-                                "type": "number"
-                            }
-                        ],
-                        "maxItems": 2,
-                        "minItems": 2
-                    }
-                }
-            },
-            "required": ["type", "count", "sum"]
-        }
-    ]
+        "unit": {
+            "type": ["string", "null"]
+        },
+        "value": {"type": "number"}
+    },
+    "required": ["type", "value"]
                 }
             },
             "additionalProperties": false