--- a/processor/metric/package_tests/processor_test.go
+++ b/processor/metric/package_tests/processor_test.go
@@ -31,6 +31,7 @@ var (
 		{Name: "TestProcessMetricMinimal", Path: "../testdata/metric/minimal.json"},
 		{Name: "TestProcessMetricMultipleSamples", Path: "../testdata/metric/multiple-samples.json"},
 		{Name: "TestProcessMetricNull", Path: "../testdata/metric/null.json"},
+		{Name: "TestProcessMetricMetricbeat", Path: "../testdata/metric/metricbeat.json"},
 	}
 )