… docs/data/intake-api/generated/metric/payload.json
… docs/spec/metrics/sample.json
… include/fields.go
… model/metric/_meta/fields.yml
… model/metric/system/_meta/fields.yml
… processor/metric/package_tests/TestProcessMetric.approved.json
… processor/metric/package_tests/TestProcessMetricMetricbeat.approved.json
… processor/metric/package_tests/TestProcessMetricMultipleSamples.approved.json
… processor/metric/package_tests/TestProcessMetricNull.approved.json
… processor/metric/system/_meta/fields.yml
… testdata/metric/metricbeat.json
… testdata/metric/minimal.json
… testdata/metric/multiple-samples.json
… testdata/metric/null.json
… testdata/metric/payload.json