In order to promote reuse of existing queries and visualizations and
possible future curated UIs