@graphaelli 2018-07-16
Added another commit to simplify the implementation now that we've scaled back our ambition.  Some unsurprising benchmark improvements as a result:

```
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkProcessor/TestProcessMetricValidate-8                          85297         65487         -23.22%
BenchmarkProcessor/TestProcessMetricDecode-8                            9276          6829          -26.38%
BenchmarkProcessor/TestProcessMetricTransform-8                         14118         10276         -27.21%
BenchmarkProcessor/TestProcessMetricProcessRequest-8                    108431        78837         -27.29%

benchmark                                                               old allocs     new allocs     delta
BenchmarkProcessor/TestProcessMetricValidate-8                          428            284            -33.64%
BenchmarkProcessor/TestProcessMetricDecode-8                            41             31             -24.39%
BenchmarkProcessor/TestProcessMetricTransform-8                         98             59             -39.80%
BenchmarkProcessor/TestProcessMetricProcessRequest-8                    567            374            -34.04%

benchmark                                                               old bytes     new bytes     delta
BenchmarkProcessor/TestProcessMetricValidate-8                          39805         29281         -26.44%
BenchmarkProcessor/TestProcessMetricDecode-8                            1480          984           -33.51%
BenchmarkProcessor/TestProcessMetricTransform-8                         8178          6406          -21.67%
BenchmarkProcessor/TestProcessMetricProcessRequest-8                    49464         36676         -25.85%
```