@graphaelli 2018-07-19
All feedback has been addressed, thanks for sticking with this as the requirements changed a few times since first opening.

I'm planning to follow up on this issue with documentation in #1060 about being careful with metric naming as 1. fields explosion could be a real issue here 2. since fields are on the root of these docs now (not under `metrics` like originally implemented), care needs to be take to avoid conflicts with apm document keys, especially when upgrading.